tsify linting configs
