--- lint-staged.config.js
@@ -1,7 +1,0 @@
-const config = {
-  '*.js': ['xo --fix'],
-  '*.{html,json}': ['prettier --write'],
-  '*.md': ['markdownlint --fix', 'prettier --write'],
-};
-
-export default config;
